QT6: Added support for QT6
* tested with Qt 5.15.2 and Qt 6.2.3
* tested on Windows (MSVC 2022) and Linux (gcc 11)

* replaced QRegExp with QRegularExpression
* replaced qSort with std::sort
* replaced QString!=0 with QString!=nullptr
* replaced QByteArray.append(QString) with QByteArray.append(QByteArray)
* replaced sprintf() with asprintf()
* replaced load() with loadRelaxed()
* replaced store() with storeRelaxed()
* added ifdefs for Qt < 5.14

+++ plugin/dltdbusplugin/dltdbusplugin.h
@@ -66,7 +66,7 @@ inline bool operator==(const DltDbusMethodKey &e1, const DltDbusMethodKey &e2)
            && e1.getSerial() == e2.getSerial();
 }
 
-inline uint qHash(const DltDbusMethodKey &key)
+inline size_t qHash(const DltDbusMethodKey &key)
 {
     return qHash(key.getSender()) ^ key.getSerial();
 }
